Split Installer class (#3)
